refactor: reorganize src folder structure
